requirements.txt
pytest-homeassistant-custom-component==0.13.320
python-dateutil>=2.8.0
